Reduce Memory Usage with FileStream instead of MemoryStream

+++ SecurityCamera.Domain/ObjectDetectionDomain/ImageDetection.cs
@@ -7,7 +7,7 @@ public class ImageDetection : IEntity
     public Guid Id { get; set; }
     public required string CameraName { get; set; }
     public required string ImageName { get; set; }
-    public required int ImageSize { get; set; }
+    public required long ImageSize { get; set; }
     public required DetectionType DetectionType { get; set; }
     public required DateTime DetectionDateTime { get; set; }
     public required string RemoteStorageContainer { get; set; }
